requirements.txt
edgartools>=5.17.1
pandas>=2.0.0